requirements.txt
virtualenv-16.6.0
selenium-3.141.0 
urllib3-1.25.3